requirements.txt
opencv-python
numpy
tensorflow
torch
torchvision
torchaudio
facenet-pytorch
mediapipe
transformers
pydub
pyannote.audio
librosa
accelerate
scikit-learn
pandas
matplotlib
moviepy
Pillow
seaborn
gradio